extracted interface off ServerService, and eliminated duplicated code
