+++ SSServer/src/main/java/com/prose/crhen/SSServer/MockDataInitializer.java
@@ -1,9 +1,8 @@
 package com.prose.crhen.SSServer;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.prose.crhen.SSServer.business.ServerService;
+import com.prose.crhen.SSServer.business.api.ServerService;
 import com.prose.crhen.SSServer.dto.ServerUpdateDTO;
 import com.prose.crhen.SSServer.dto.VolumesUpdateDTO;
 import org.slf4j.Logger;
