+++ SSServer/src/main/resources/application.properties
@@ -1,2 +1,2 @@
 server.port = 9010
-spring.profiles.active=test,mockdata
+spring.profiles.active=dev,mockdata,prod
